LayoutTests/imported/w3c/web-platform-tests/css/css-tables/bounding-box-computation-1-expected.txt

Bounding box computation

Checks that all table-internal boxes are positioned correctly

The all table-internal elements of this table shoud be 100px and 100px wide tall:
Table size is 120px. There are two 10px border-spacing, one at each side.


PASS Control test: Table width is 120px
PASS Control test: Table height is 120px
PASS Table-cell is 100px wide
PASS Table-cell is 100px tall
FAIL Table-row is 100px wide assert_equals: expected 100 but got 120
PASS Table-row is 100px tall
FAIL Table-row-group is 100px wide assert_equals: expected 100 but got 120
FAIL Table-row-group is 100px tall assert_equals: expected 100 but got 120
PASS Table-column is 100px wide
FAIL Table-column is 100px tall assert_equals: expected 100 but got 120
PASS Table-column-group is 100px wide
FAIL Table-column-group is 100px tall assert_equals: expected 100 but got 120
